requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.7
django-cleanup==8.1.0
django-crispy-forms==2.2
pillow==10.4.0
python-vlc==3.0.20123
sqlparse==0.5.0
typing_extensions==4.12.2
requests==2.32.3
RPi.GPIO==0.7.1
django-environ==0.11.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
pytz==2024.1
